semi/requirements.txt
imapclient>=3.0.1
scikit-learn>=1.3.0
numpy>=1.24.0
email-validator>=2.0.0
python-dotenv>=1.0.0
click>=8.1.0
colorama>=0.4.6
flask>=2.3.0
flask-cors>=4.0.0
google-auth>=2.21.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.95.0
gunicorn>=21.2.0